backend/requirements.txt
fastapi==0.95.2
fastapi-mcp==0.1.4
uvicorn[standard]==0.22.0
pyyaml==6.0
